Expose `NavigationGeocodedPlacemark` as a separate file. Improve documentation.
